# infra/openapi/novax-mvp.yaml
openapi: 3.1.0
info:
  title: NovaX MVP API
  version: 0.1.0
  description: Initial contract for frontend-backend integration across gateway, auth, catalog, cart and order flows.
servers:
  - url: http://localhost:8080/api/v1
paths:
  /health:
    get:
      summary: API Gateway health
      operationId: gatewayHealth
      responses:
        '200':
          description: Service health status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
  /auth/login:
    post:
      summary: User login
      operationId: authLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthTokens'
  /products:
    get:
      summary: List products with filter and pagination
      operationId: listProducts
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: minPrice
          schema:
            type: integer
        - in: query
          name: maxPrice
          schema:
            type: integer
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
  /products/{slug}:
    get:
      summary: Product detail
      operationId: getProductDetail
      parameters:
        - in: path
          name: slug
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
  /cart:
    get:
      summary: Get authenticated user's cart
      operationId: getCart
      responses:
        '200':
          description: Cart detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
    put:
      summary: Replace cart items
      operationId: upsertCart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CartUpsertRequest'
      responses:
        '200':
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CartResponse'
  /orders:
    post:
      summary: Create order from cart
      operationId: createOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Created order
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
  /ai/recommendations:
    post:
      summary: AI recommendations based on intent and catalog context
      operationId: recommendProducts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AiRecommendationRequest'
      responses:
        '200':
          description: Recommended products
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'
components:
  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          examples: [ok]
        service:
          type: string
        timestamp:
          type: string
          format: date-time
      required: [status, service, timestamp]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AuthTokens:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
      required: [accessToken, refreshToken]
    Product:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: integer
        originalPrice:
          type: integer
        rating:
          type: number
        image:
          type: string
          format: uri
        category:
          type: string
      required: [id, slug, name, price]
    ProductListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
      required: [items, page, pageSize, total]
    CartItem:
      type: object
      properties:
        productId:
          type: string
        quantity:
          type: integer
          minimum: 1
      required: [productId, quantity]
    CartResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
        subtotal:
          type: integer
        total:
          type: integer
      required: [items, subtotal, total]
    CartUpsertRequest:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/CartItem'
      required: [items]
    CreateOrderRequest:
      type: object
      properties:
        shippingAddressId:
          type: string
        paymentMethod:
          type: string
          enum: [cod, vnpay, momo]
        note:
          type: string
      required: [shippingAddressId, paymentMethod]
    Order:
      type: object
      properties:
        id:
          type: string
        orderNumber:
          type: string
        status:
          type: string
          enum: [pending, confirmed, processing, shipped, delivered, cancelled]
        total:
          type: integer
      required: [id, orderNumber, status, total]
    AiRecommendationRequest:
      type: object
      properties:
        intent:
          type: string
        budgetMin:
          type: integer
        budgetMax:
          type: integer
        preferredBrands:
          type: array
          items:
            type: string
      required: [intent]